+++ Craftsman/Services/FileNames.cs
@@ -10,6 +10,16 @@ public static class FileNames
     {
         return "MassTransitServiceExtension";
     }
+    
+    public static string MessageClassName(string messageName)
+    {
+        return $"{messageName}";
+    }
+    
+    public static string MessageInterfaceName(string messageName)
+    {
+        return $"I{messageName}";
+    }
 
     public static string EntityCreatedDomainMessage(string entityName)
     {